VERSION: proposed new .so version numbers
This commit reflects an update in the Libtool version numbers for many
of the major shared libraries that Open MPI produces.  These version
numbers reflect the first in the v1.10.x series, and are designed to
*not* be backwards compatible with the v1.8.x series.  As such, most
of the version numbers simply went from x:y:z to (x+5):0:0.  This
accomplishes two things:

1. Makes the new version be non-backwards-compatible (in an ABI sense)
2. Leaves a "gap" in the "c"urrent value for future 1.8.x releases (if
   we need them)

Fixes open-mpi/ompi#734.