Java bindings for window name get and set
Includes bindings for MPI_WIN_GET_NAME and MPI_WIN_SET_NAME. Rather
than call into the C code to set the window name using MPI calls,
we decided it would be better to implement purely in the Java.
The synchronized keyword is used on the get and set methods to ensure
thread safety.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>