Makefile.am's: if calling OPAL functions, must link to it
On some OSs (e.g., Ubuntu 14.04.2 LTS), the linker is configured such
that the symbols of library dependencies are not available to the
application.  Hence, we need to explicitly list such dependencies when
creating the executable.

For this commit, these tests are use OPAL function calls, so we must
explicitly link in libopen-pal.so.