fortran mpifh: only compile the sizeof lib when necessary
This change was already on master; it was apparently made a long time
ago.  Hence, this commit is not a cherry-pick from master (I didn't
go back to try to find that specific fix); it's a unique commit on
v1.10 to bring this specific section of mpif-h/Makefile.am in-line
with what is on master.

Namely: only create the libmpi_mpifh_sizeof.la if there is a need to.
Otherwise, use AM_CONDITIONAL logic to not build it at all.