opal: add recursive mutex
This new class is the same as the opal_mutex_t class but has a
different constructor. This constructor adds the recursive flag to the
mutex attributes for the lock. This class can be used where there may
be re-enty into the lock from within the same thread.

master commit open-mpi/ompi@54998e5745424a683fec8653831c85a5dc8ac9b5

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>