mca/base: add priority output to mca_base_select
The mca_base_select function uses returned priorities to select the
best component/module. This priority may be of use to the caller so
pass that information back in an optional argument. If the priority is
not needed pass NULL.

This is a modification of
open-mpi/ompi@8b5810f7f7cbaadcd3eba528404b4c417dc0e44e . The primary
difference is we left the original mca_base_select in place for
compatibility.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>