btl/openib: use only SRQ on ib by default
It was decided some time ago that there is no benefit to using any
per-peer receive queues on infiniband. At the time we decided not to
change the default but that objection has been dropped. This commit
changes the 128 message queue to use SRQ instead of PP. This has no
impact on iWarp which sets the default in a different way.

Closes open-mpi/ompi#1156

(cherry picked from open-mpi/ompi@b24b3a4ae467b7299b42075ab77981a92532427f)

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>