travis: use "make -k" to see all failures
One of the criticisms of Travis on the call earlier today was that
when there's a compile failure, you really only see the *first*
compile failure.  Then you fix that, resubmit, and 15-60 minutes
later, you see the *next* compile failure.

Per some discussion with @bosilca and @rhc54 this afternoon, use "make
-k", which will have "make" try to continue the build even after
compile failures.  Hence, you'll be able to see many more compile
errors that just one at a time.

In some cases (e.g., if there's a typo in a top-level header file), it
will cause a bazillion lines of compile error output, but hopefully
that should be easy to spot / ignore.