VERSION: update .so version numbers for v2.0.0
The highest .so "current" version number in the v1.10 series (as of 18
Dec 2015, v1.10.1) is 13.  We don't want Open MPI v2.0.0 to be
backwards compatible with v1.10, so we need to update "current" to be
a higher value, and set "age" to be "0".

But the v1.10 series has a little life left in it, so we want to leave
a gap between the highest "current" value for v1.10 and where we start
v2.0.0.  So -- let's add 5 to 13 and start with 18.  ...and then let's
round up to 20, just because it's nice and similar to 2.0.0.  :-)

Finally, we might as well update *all* the .so version numbers to
20:0:0 and start them all off at the same level.

If we're lucky, we can update "current" to 30 for v3.0.0.  :-)

Fixes open-mpi/ompi#1159