VERSION: update .so version numbers for v1.10.2
This commit amends ce112382893e13d9f0d9d0746e0c434c74fdf114, which was
an attempt at updating the .so version numbers post-1.10.1.  This
commit completes that commit (because we're much closer to the v1.10.2
release) and fixes the incorrect update of the oshmem .so version
number from that commit.