btl/ugni: fix race condition when adding endpoint to wait list
This commit fixes a race condition that can cause an endpoint to be
added to the wait list multiple times. To fix the issue an additional
check has been added to ensure the endpoint is not on the wait list
after the wait list lock is held. The wait list processing code has
also been updated to keep the wait list lock until all wait listed
endpoints have been handled. This reduces the chance that an endpoint
that is being processed by the wait list code is not re-added to the
list by a competing send.

(cherry picked from open-mpi/ompi@14704201e26a5e693c4b16a19c15c3fdd84dc18f)

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>