opal.pc.in: fix typo; use the write AC_SUBST'ed variable
As reported by @marksantcroos, this substitution in opal.pc was
incorrect -- it left @{libdir} in the string (vs. ${libdir}).  The fix
is simple: use the proper substitution variable in opal.pc (it was
never updated to reflect the new/correct name that was created just
for the pkg-config files).

Fixes open-mpi/ompi#1343.