ompi: temporarily disable dynamic add_procs on big-endian and 32-bit
This commit temporarily disables the dynamic add_procs support on
big-endian and 32-bit systems. There is an issue with the sentinel
used to indicate when proc needs to be generated that causes issues on
these systems. This will be removed once the underlying issue has been
resolved.

Fixes open-mpi/ompi#1348

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>