osc/pt2pt: do no use OPAL_THREAD_ADD64 for lock serial number
Some platforms do no support 64-bit atomic math. osc/pt2pt was using
OPAL_THREAD_ADD64 unconditionally which was causing compilation to
fail on these platforms. The function in question was alreay grabbing
a lock on the osc/pt2pt module so moved the increment to after the
lock to protect the module member (instead of using an atomic).

Fixes open-mpi/ompi#1162

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>